requirements.txt
alembic>=1.13.1
Authlib>=1.6.5
certifi>=2025.11.12
configobj>=5.0.9

# Azure SDK packages (optional, for Azure Key Vault integration)
azure-identity>=1.25.1
azure-keyvault-secrets>=4.7.0
azure-storage-blob>=12.27.1
azure-storage-file-share>=12.23.1
azure-storage-queue>=12.14.1

# OpenBao/HashiCorp Vault integration (optional, for self-hosted secrets management)
hvac>=2.4.0
backoff>=2.2.1
beautifulsoup4>=4.14.3
bleach>=6.1.0
Brotli>=1.2.0
celery>=5.6.0
click>=8.1.7
cryptography>=46.0.3
dnspython>=2.8.0
email-validator>=2.3.0
filelock>=3.20.1
Flask>=3.1.2
Flask_Bcrypt>=1.0.1
Flask_Cors>=6.0.1
Flask_Limiter>=4.1.1
flask_mail>=0.10.0
Flask_Migrate>=4.0.7
Flask_Principal>=0.4.0
flask_sqlalchemy>=3.1.1
flask_wtf>=1.2.2
fonttools>=4.60.2
geoip2>=4.8.0
idna>=3.7
jinja2>=3.1.6
jsonschema>=4.25.0
langchain>=1.2.0
langchain_core>=1.2.5
langchain_openai>=1.1.6
llama_index>=0.14.12
lmdb>=1.6.2
logzero>=1.7.0
Markdown>=3.7
MarkupSafe>=3.0.3
marshmallow>=3.26.2
openai>=1.109.1
passlib>=1.7.4
pillow>=12.1.0
pip>=25.3
psutil>=7.0.0
psycopg2_binary>=2.9.9
pydantic>=2.12.5
pypdf>=6.5.0
pytest>=8.3.5
python-dotenv>=1.2.1
python-socketio>=5.15.0
qrcode>=8.2
redis>=7.1.0
Requests>=2.32.5
setuptools>=80.9.0
SQLAlchemy>=2.0.45
stripe>=14.0.1
tiktoken>=0.12.0
twisted>=24.7.0
urllib3>=2.6.2
weasyprint>=67.0
tinycss2>=1.5.1
websocket_client>=1.9.0
Werkzeug>=3.1.4
WTForms>=3.2.1
zipp>=3.19.1
nats-py>=2.11.0
